Remove duplicate code in sagemaker/core/training
