+++ sagemaker-train/src/sagemaker/train/constants.py
@@ -10,15 +10,11 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
-"""
-DEPRECATED: This module has been moved to sagemaker.core.training.constants
-
-This is a backward compatibility shim. Please update your imports to:
-    from sagemaker.core.training.constants import ...
-"""
+"""Constants module."""
 from __future__ import absolute_import
-
 import os
+
+DEFAULT_INSTANCE_TYPE = "ml.m5.xlarge"
 
 SM_CODE = "code"
 SM_CODE_CONTAINER_PATH = "/opt/ml/input/data/code"
