+++ sagemaker-train/tests/unit/test_training_constants.py
@@ -10,11 +10,11 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
-"""Unit tests for sagemaker.core.training.constants module."""
+"""Unit tests for sagemaker.train.constants module."""
 from __future__ import absolute_import
 
 import os
-from sagemaker.core.training import constants
+from sagemaker.train import constants
 
 
 class TestTrainingConstants:
